+++ android/app/src/main/java/com/xianxia/sect/core/engine/service/DiplomacyService.kt
@@ -8,6 +8,7 @@ import com.xianxia.sect.core.GameConfig
 import com.xianxia.sect.core.state.GameStateStore
 import com.xianxia.sect.core.state.MutableGameState
 import com.xianxia.sect.di.ApplicationScopeProvider
+import com.xianxia.sect.core.engine.system.DiplomacySystem
 import com.xianxia.sect.core.engine.system.GameSystem
 import com.xianxia.sect.core.engine.system.AddResult
 import com.xianxia.sect.core.engine.system.InventorySystem